Commit from Kit's Zotac (new hip controller stuff).
git-svn-id: https://atrias.googlecode.com/svn/trunk@569 a504bc44-2baa-4cb8-4136-ba805e94d278

+++ drl-sim/drl_library/src/discretize.cpp
@@ -5,7 +5,7 @@ uint16_t discretize(float val, float min, float max)
 	return (uint16_t)((val - min) / (max - min) * MAX_16BIT_UINT);
 }
 
-float undiscretize(uint16_t dis, float min, float max)
+float undiscretize(uint32_t dis, float min, float max)
 {
 	return (float)dis / (float)MAX_16BIT_UINT * (max - min) + min;
 }
